Avoid serialization of anonymous classes
